refactor(container): mv CommunityBanner -> CommunityDigest
