--- src/containers/banner/CommunityBanner/tests/store.test.js
@@ -1,10 +1,0 @@
-/*
- * CommunityBanner store test
- *
- */
-
-// import CommunityBanner from '../index'
-
-it('TODO: store test CommunityBanner', () => {
-  expect(1 + 1).toBe(2)
-})
