+++ src/containers/digest/CommunityDigest/BriefView.js
