+++ src/containers/digest/CommunityDigest/DigestView/ColumnView/index.js
