+++ src/containers/digest/CommunityDigest/DigestView/RowView/TeamList.js
