+++ src/containers/digest/CommunityDigest/DigestView/RowView/index.js
