+++ src/containers/digest/CommunityDigest/DigestView/index.js
