+++ src/containers/digest/CommunityDigest/SocialList.js
