+++ src/containers/digest/CommunityDigest/schema.js
