+++ src/containers/digest/CommunityDigest/styles/brief_view.js
