+++ src/containers/digest/CommunityDigest/styles/digest_view/column_view/index.js
