+++ src/containers/digest/CommunityDigest/styles/digest_view/row_view/team_list.js
