+++ src/containers/digest/CommunityDigest/styles/expand_texts.js
