+++ src/containers/digest/CommunityDigest/styles/social_list.js
