+++ src/containers/digest/CommunityDigest/tests/store.test.js
@@ -1,0 +1,10 @@
+/*
+ * CommunityDigest store test
+ *
+ */
+
+// import CommunityDigest from '../index'
+
+it('TODO: store test CommunityDigest', () => {
+  expect(1 + 1).toBe(2)
+})
