+++ src/pages/home/posts.js
@@ -24,7 +24,7 @@ import {
 } from '@/utils'
 
 import GlobalLayout from '@/containers/GlobalLayout'
-import CommunityBanner from '@/containers/banner/CommunityBanner'
+import CommunityDigest from '@/containers/digest/CommunityDigest'
 import CommunityContent from '@/containers/content/CommunityContent'
 
 import { P } from '@/schemas'
@@ -172,7 +172,7 @@ const CommunityPage = (props) => {
         errorCode={errorCode}
         errorPath={community.raw}
       >
-        <CommunityBanner />
+        <CommunityDigest />
         <CommunityContent />
       </GlobalLayout>
     </Provider>
